used shared enums and adjusted AddressIF
