--- src/components/base-address/enums/index.ts
@@ -1,1 +1,0 @@
-export * from './validation-rule'
