--- src/components/base-address/enums/validation-rule.ts
@@ -1,9 +1,0 @@
-export enum ValidationRule {
-  BC = 'BC',
-  CANADA = 'Canada',
-  MAX_LENGTH = 'maxLength',
-  MIN_LENGTH = 'minLength',
-  POSTAL_CODE = 'postalCode',
-  REQUIRED = 'required',
-  ZIP_CODE = 'zipCode'
-}
