+++ src/enums/address-validation-rules.ts
@@ -1,0 +1,9 @@
+export enum ValidationRule {
+    BC = 'BC',
+    CANADA = 'Canada',
+    MAX_LENGTH = 'maxLength',
+    MIN_LENGTH = 'minLength',
+    POSTAL_CODE = 'postalCode',
+    REQUIRED = 'required',
+    ZIP_CODE = 'zipCode'
+  }
